Added caching logic and spreadsheet column append
The choice for now is to make a copy spreadsheet to avoid damage to the orginial data. However, the new column has been appended to the dataframe, so it can use it from then on directly. Later on better control logic is recommended.

However, this should be enough to finish Naive Bayes.